feat(theme-builder): tailwind: move withPrintMode() to a dedicated file
