+++ packages/theme-builder/src/tailwind/index.ts
@@ -3,3 +3,8 @@ export {
 
   defaultColors,
 } from './common';
+
+export {
+  darkStyleNotPrintPlugin,
+  withPrintMode,
+} from './plugin-print-mode';
